// interfaces/libros/ILibro.ts
// Generated by https://quicktype.io

export interface ILibro {
    id:               string;
    title:            string;
    isbn:             string;
    pageCount:        number;
    publishedDate:    string;
    thumbnailUrl:     string;
    shortDescription: string;
    longDescription:  string;
    status:           string;
    precio:           number;
}